refactor: flatten service interface and impl
The service implementations are not included in X-Ray tracing due to the
`xServiceImpl` naming not fitting the configured `*Service` wildcard.
Rather than change the wildcard, the services should be flattened to a
single `xService` instead of an interface and implemenation class.
There are no alternative implementations of the service so there is no
need to overcomplicate with an extracted interface.

TIS21-SHED
