+++ build.gradle
@@ -10,7 +10,7 @@ plugins {
 }
 
 group = "uk.nhs.hee.tis.trainee"
-version = "0.8.0"
+version = "0.8.1"
 
 configurations {
   compileOnly {
